code/requirements.txt
google-generativeai>=0.8.0
chromadb>=0.4.22
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0

